besouro automatic message

+++ src/org/univoulu/tol/sqatlab/sudoku/SudokuVerifier.java
@@ -10,7 +10,7 @@ public class SudokuVerifier {
 		Return -4: means it is violating Rule #4
 		return -1 if the length is too short*/
 		
-		System.out.println(Integer.parseInt(candidateSolution));
+		System.out.println(Long.parseLong(candidateSolution));
 		
 		if(Integer.parseInt(candidateSolution) < 0){
